DailyOddsEnv: change lower bound of observation space to 0
In DailyOddsEnv, we pad the odds with zeros if there are less games
than in the day with the maximum number of games in it (bounded by
number of teams divided by 2). That causes the get_odds method to return
observations that do not belong to the defined observation space.