Reduce dimensionality of action spaces and change action representation
Change representation of the action (the integer representing which odds
to choose) to a binary representation, where each odd in the action is "1"
and each odd that isn't is "0".
Reduce dimensionality of Percentage environments by deducing the action
itself by the given betting percentages for each outcome (creating the
binary representation and deducing the action from it).